+++ examples/__init__.py
@@ -1,0 +1,4 @@
+# SPDX-FileCopyrightText: 2026 IonQ, Inc.
+# SPDX-License-Identifier: Apache-2.0
+
+"""Runnable ionq-core examples."""